数据库delete方法支持调用useSoftDelete方法进行软删除

+++ library/think/db/Query.php
@@ -2139,6 +2139,20 @@ class Query
     {
         $this->parseOptions();
 
+        if (!empty($this->options['soft_delete'])) {
+            // 软删除
+            if (!is_null($data) && true !== $data) {
+                // AR模式分析主键条件
+                $this->parsePkWhere($data);
+            }
+
+            list($field, $condition) = $this->options['soft_delete'];
+            unset($this->options['soft_delete']);
+            $this->options['data'] = [$field => $condition];
+
+            return $this->connection->update($this);
+        }
+
         $this->options['data'] = $data;
 
         return $this->connection->delete($this);
